telegram-bot/requirements.txt
# Telegram Bot Dependencies
python-telegram-bot==20.7
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
celery==5.3.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==7.0.0